requirements.txt
absl-py==0.12.0
amqp==5.0.5
asgiref==3.3.1
astunparse==1.6.3
beautifulsoup4==4.9.3
billiard==3.6.3.0
cached-property==1.5.2
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
eventlet==0.30.2
flatbuffers==1.12
gast==0.3.3
ghettoq==0.4.5
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.1
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kombu==5.0.2
Markdown==3.3.4
nltk==3.6.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pickle-mixin==1.0.2
prompt-toolkit==3.0.14
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.2.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
soupsieve==2.1
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
